refactor(network): move `Compression` and `Algorithms` into separate modules, get rid of `CompressionAlgorithm` type
